Move SBRController related code to its own file
